data/requirements.txt
boto3==1.35.97
botocore==1.35.97
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
urllib3==2.3.0
